gcms-data-platform/streamlit_app/requirements.txt
streamlit==1.29.0
snowflake-connector-python==3.6.0
pandas==2.1.4
cryptography==41.0.7
plotly==5.18.0
